Allow recording of activities
Allows users to record activities for today. This is the most basic
feature of the application. As of now you cannot change the date of an
activity or record an activity for a day other than today.

Note: The generator does not pick up on the existing activity schema
when generating so I'll have to add all of the forms etc manually.